fp_updater: Change bloonchipper firmware used for testing
The firmware used for testing FPMCU updates doesn't support
communication with AP using UART. As a result it wasn't possible to
update firmware because there was no communication with FPMCU after it
jumped to RW part.

The test will be still failing on Zork and Guybrush because biod is not
disabled before starting the test, so 'bio_wash' called after running
'flash_fp_mcu' will fail.

The key used to sign both bloonchipper binaries (with RO
bloonchipper_v2.0.4277 and bloonchipper_v2.0.5938) is the same, and
during the test we flash only RW part, so one bloonchipper firmware can
be used to test boards with SPI and UART communication interface.

BUG=b:239422436
TEST=Run 'tast run <IP> firmware.FpUpdater' on Hatch. Make sure test
     passes.
TEST=Run 'tast run <IP> firmware.FpUpdater' on Zork. Make sure test
     doesn't fail on 'Updater did not succeed, please check output dir'.

Change-Id: Ibf3f52208d4b18656a5e23338492f196f43fedbe
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3768407
Tested-by: Patryk Duda <patrykd@google.com>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Patryk Duda <patrykd@google.com>
Reviewed-by: Mohamed Omar <mohamedaomar@google.com>
Reviewed-by: Tom Hughes <tomhughes@chromium.org>