video: Add test expectations helper functions
Most video decode test cases pass only if an entire set of test vectors
are able to decode successfully. If one file in the set is failing for
a known, tracked reason, having the test fail blocks a test triager's
ability to detect regressions.

This CL adds functions and definitions for loading a test expectation
file from disk. This gives a test writer the ability to allow a test to
pass based on an expectation that a particular file fails to decode on
a particular device. The test expectation file is YAML, but the
structure of the file is to be defined by the needs of a test case. The
path is determined by the test name and the device type.

E.g. a trogdor device may have a PlatformDecoding test expectation at:

/usr/local/graphics/expectations/tast/video/PlatformDecoding/platform-Trogdor.yml

The device may match a single test file by its model, board, or platform
(queried in that order.)

BUG=b:230389196
TEST=tast run <DUT> video.PlatformDecoding.*

Change-Id: Ie87faf45de5db7da0dbefb2e037da5cfaa3ae67e