video: Add test expectations helper functions
Most video decode test cases pass only if an entire set of test vectors
are able to decode successfully. If one file in the set is failing for
a known, tracked reason, having the test fail blocks a test triager's
ability to detect regressions.

This CL adds functions and definitions for loading a test expectation
file from disk. This gives a test writer the ability to allow a test to
pass based on an expectation that a particular file fails to decode on
a particular device. The test expectation file is YAML. The path is
determined by the test name and the device type.

E.g. a trogdor device may have a PlatformDecoding test expectation at:

/usr/local/graphics/expectations/tast/video/PlatformDecoding/board-trogdor.yml

The device may match a single test file by its model, build, board, or
gpu chipset (queried in that order.)

The YAML structure is a map from the test name to an
expectations.Expectation structure. I.e.

video.PlatformDecoding.v4l2_vp8_comprehensive
  expectation: FAIL
  tickets: ["b/12345", "crbug/67890"]
  comments: "Notes from the triager"
  sinceBuild: "R100-14526.89.0"

A YAML for a parameterized tests may contain up to one key per test
case. For a non-parameterized test, there will be at most one key. Test
cases without an expectation are expected to pass.

To use test expectations, the test writer should follow the guide in
expectations/README.md

BUG=b:230389196
TEST=tast run <DUT> video.PlatformDecoding.* positive and negative
TEST= testing with and without expectations files

Cq-Depend: chromium:3843102
Change-Id: Ie87faf45de5db7da0dbefb2e037da5cfaa3ae67e