video: Add PlatformDecoding test expectations
PlatformDecoding uses a test-specified decoder binary to decode a set of
test vectors. If a single test vector in the set fails to decode
correctly, then the test fails. This impacts the ability to detect
regressions in the other vectors in the test case set.

This CL adds the ability to load and rewrite test expectations files for
PlatformDecoding test cases.

The test expectations file of the file is YAML with the format:

<decoder file name>:
  <test vector base name one>:
    expectation: FAILURE
    ticket: b/230389196
  <test vector base name two>:
    expectation: FAILURE
    ticket: b/230389196

If a test vector is not specified, it is expected that it should pass.

To update test expectations for a test case, use tast variables. To
update the expectations file:
video.PlatformDecoding.updateExpectations=true

To change the file type do match by model, board, or platform, use:
video.PlatformDecoding.expectationsFileType=model|board|platform

To specify a ticket for newly failing test cases, use:
video.PlatformDecoding.defaultFailureTicket="b/1234567"

BUG=b:230389196
TEST=tast run <DUT> video.PlatformDecoding.*

Change-Id: I5927def04996a0b6c276f6ba51e4cfdc916bc0ca