tast-tests: Add StartReboot to router controllers
This adds a new method, StartReboot, to the base router controller
interface which is intended to start a system reboot on the
controlled router when called. This method has been implemented
for all router controllers, but only intended to be used with
OpenWrt routers. The Legacy and AX router controllers will
return an error if it is called as of now, but this can be changed
later if it is later allowed to reboot those as well between tests.

BUG=b:211517577
TEST=TODO

Change-Id: I52abbe8fad1d28139a0bcadbded60e6ac7d6ffbd