tast-tests: Reboot OpenWrt routers during Reinit called at PreTest
This change includes a few key things to practically reboot Openwrt
routers between tests with minimal changes to the current way routers
are controlled.

First, TestFixture.routers has been refactored to a list of pointers
in order to allow for easier access to each object. The items were
already being used each time through manually created pointers, so
this is more of just a simplification for managing changing routerData.

Second, the TestFixture.Reinit function has been updated to also
reboot OpenWrt routers. To reboot the router and still maintain its
control over it with a router controller instance, it will trigger
a router reboot, wait a few minutes, reconnect with a fresh ssh
connection, and then replace the old router controller instance with
a new instance using the new ssh connection.

Third, the TestFixture.Reinit function is now called during PreTest,
instead of during Reset, so that tests start immediately after a
fresh reboot when using OpenWrt routers.

Finally, relevant timeouts have been updated to account for the
additional time it takes for routers to be rebooted between
tests.

BUG=b:211517577
TEST=Successfully ran wifi.SimpleConnect.80211a with a Legacy router
and confirmed that the router was not rebooted. Ran all
of wifi.SimpleConnect.* with an OpenWrt router and confirmed that the
router was rebooted between tests and that tests were still able to
perform the same way after a tast-initiated reboot as with a manual
reboot.

Change-Id: I4b8a8d0e845a14cb378510c008fe1ffe5393a074
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3771027
Tested-by: Jared Bennett <jaredbennett@google.com>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Commit-Queue: Jared Bennett <jaredbennett@google.com>