tast-tests: Use unique working directory with OpenWrt routers
Since we are now recreating OpenWrt router controller instances
as a result of rebooting the router, it is more likely that we'll
have two router controller instances active at the same time for
the same router. In order to ensure that their working files do
not interfere with each other (e.g. we don't want one's Close to
delete files used by an active router controller), the working
directory on the router has been moved to a subdir under the
previous working directory location. This subdir is named with a
human-readable timestamp of the current time during router
controller instance creation so that it will always be unique
between different instances.

BUG=b:211517577
TEST=Successfully ran wifi.SimpleConnect.80211a with an OpenWrt
router. Confirmed during the test by checking the router that
the new working directory was created and used as expected.

Change-Id: I47a7e057252998575ac92aa6be3c3d7d26204d4e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3771030
Commit-Queue: Jared Bennett <jaredbennett@google.com>
Reviewed-by: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: Patrick Meiring <meiring@google.com>
Reviewed-by: Kevin Lund <kglund@google.com>
Tested-by: Jared Bennett <jaredbennett@google.com>