tast-tests: More complete Hostapd WEP config check
Not all Hostapd WEP configurations were caught by the previous
check. This adds support for the checking the rest of the known WEP
configurations. Since this check is not specific to a router type
and the logic was getting more complex, it has been moved to a common
function usable by all router types, not just OpenWrt.

BUG=b:211517577
TEST=Successfully passed wifi.SimpleConnect.80211a with an OpenWrt
router. Successfully failed wifi.SimpleConnect subtests that use
WEP (8021xwep, wep104, wep40, wephidden) at the WEP check with
an OpenWrt router.

Change-Id: I2fc1d4fcc9717f76c5f1268977122e7efb59de82