tast-tests: Update tests to support multiple credentials update
This test adds more coverage for update credentials code paths. Now it
adds two types of credentials - pins and passwords. Updates both of them
and checks if we can still login and unlock with relevant credentials.
The updated test also ensures that we old or invalid credential is not
accepted.

BUG=b:235874130
TEST=tast run <DUT_IP> *UpdateCredentialAuthSession*

Change-Id: I713001e9e94825a7a0d577cdad4226398498ab6f