Add tast test for ARC downloads in holding space
This test uses the play store to download the beta version of Chrome
as an ARC app, then downloads a small file to verify that ARC download
show up in holding space. This test is currently expected to fail due
to crbug/1291882, but I've verified that it will pass once that bug is
resolved and the download does show properly.

BUG=b:200987303
TEST=tast run <DUT> holdingspace.ArcDownload

Change-Id: I26f5d488906d4de156dc024bfacbaf9bb7d3406d