tast-tests: Add ConnectToBTPeers helper function and test
This adds a new "ConnectToBTPeers" bluetooth fixture helper function
and test, TestConnectToBTPeers, that shows how it can be used to
connect to btpeers and send them Chameleond method calls in local
tests. In addition, the bluetooth fixture file has been moved to
the appropriate package, as it should not be in the same entry package
as the tests are.

The ConnectToBTPeers function can be used within any BT local test
as long as the "btpeers" test var is set appropriately and it is
added to the Test.Vars for that test. Reusable constants have been
added to make this simple and consistent.

Example btpeers usage:
tast run -var=btpeers=<btpeer1_ip>,<btpeer2_ip> <dut> <test>

BUG=b:217633791
TEST=tast unit tests passed. Ran test TestConnectToBTPeers
and was able to successfully connect to two btpeers as well as
execute Chameleond methods on them successfully. Was able to
call the "ResetBluetoothRef" method, though the btpeer device
used did not appear to be appropriately configured to execute
this call, so it failed.

Change-Id: I31d293142671635c8ccf2197a1679f3e2986118d