power: Add Adaptive Charging tests
This adds Adaptive Charging tests, power.AdaptiveCharging.charge_now and
power.AdaptiveCharging.settings. These are end-to-end tests for Adaptive
Charging that verify that the feature starts under the right conditions,
clicking the charge now button starts charging, and enabling/disabling
the feature in settings works correctly.

BUG=b:235902052
TEST=`tast run -var=servo=<servo hostname>:<servo port> <dut IP>
power.AdaptiveCharging` on a brya system in the lab

Change-Id: Ia74598275f5b878516783365f1d17d6002573671