tast-tests: Trace chrome in sys-wide mode
Change trace verification to a more realistic scenario when
perf doesn't launch a command and instead collects the profile from a
chrome workload.
Disable timeless decoding in system-wide since it was the issue when UI
wasn't On.
Increase ETM strobing period to make the profile smaller.
Flag the error if chrome is not in the trace.

BUG=b:225407997,b:236418174
TEST=tast run <dut> platform.PerfETM

Change-Id: I03d4ab28a1e2f944ccd6ec9d74a0f684100cfcd0