vmDLC: Tell Chrome to log in
dlcservice needs to interact with the session manager, so there's a
Chrome dependency in there. Additionally, many of these VM tests are
moderately long and/or performance-sensitive. In particular, sitting at
OOBE can generate a lot of CPU activity (sitting on video animations,
etc.) which depends on the state we're left in by prior tests.

Teaching individual tests to log in after the vmDLC fixture would also
not be simple, since the common ways to do this involve clearing some
UI/session state, which can perturb the DLCs we just installed.

Let's just log in by default, to get a consistent, quiesced Chrome
state.

This improves consistency of performance on vm.Virtiofs, so it doesn't
depend on what tests ran before it.

BUG=b:238655078
TEST=`tast run vm.Audio* vm.Virtiofs`

Change-Id: Id08bdc0a374e6a2775a05f95872a5fc3f30f6f59