arc: Add restore pre-snap bounds tests to arc.SplitView
Currently, if we snap an ARC window by dragging the caption, and close
it at the snapped state, the window stays at the edge of the screen
after reopening.
In a recent change(ag/19368231) we fixed the issue and this CL adds
the related tast tests to make sure that the pre-snap bounds can be
restored after reopening the window.

BUG=b:217609117
TEST=tast run ${DUT} arc.SplitView*

Change-Id: Idcaa08a7c582ddfe1aa4a6ad27940f05ee618860