tast-tests: Remove Downloads bind mount from nearbyshare tests
The ~/Downloads bind mount is being deprecated and thus all usages of it
need to be migrated. The canonical path is the
/home/chronos/u-HASH/MyFiles/Downloads. This required hooking up the
nearbyservice to retrieve the user's DownloadsPath and pass it back to
the remote fixtures to ensure the remote tests have access to compute
their hashes.

BUG=b:233132576
TEST=./fast_build.sh -T

Change-Id: I82889aee98674d739f088e4a6095891c01c00104
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3778837
Reviewed-by: Kyle Shimabukuro <kyleshima@chromium.org>
Tested-by: Ben Reich <benreich@chromium.org>
Commit-Queue: Ben Reich <benreich@chromium.org>