ml: Add fake adaptive charging ml service
This adds local functionality to start a fake version of the adaptive
charging ml service. This can't be part of the test that uses it, since
tast tests require only one function be exported, and the external DBus
method needs to be exported for this.

BUG=b:235902052
TEST=`tast run -var=servo=<servo hostname>:<servo port> <dut IP>
power.AdaptiveCharging` on a brya system in the lab

Change-Id: I0ce3c30f2813b0b03f4bc4dfb40daecd5f7e961f