tast-tests: Disable filemanager.SendToTrash test
Support for Drive trash is not going to make the first release, so
disable the test now to ensure it isn't running by the time Chrome
changes get landed and uprevved. The local scenario is also not
particularly useful as a tast test as the full functionality can be done
in browser tests.

BUG=b:239769339, chromium:1341867
TEST=tast list DUT '("group:mainline")' with SendToTrash excluded

Change-Id: I855731f827edf838cb9c6d57f9fb0da505549003
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3778839
Reviewed-by: Noel Gordon <noel@chromium.org>
Commit-Queue: Ben Reich <benreich@chromium.org>
Tested-by: Ben Reich <benreich@chromium.org>