meta: Add test to capture Google Docs CUJ cold boot metrics
From a practical perspective, Lacros architects are likely interested in
the time of the first Google Docs opening, not multiple opening times in
a row. The latter is what lacros.DocsCUJ measures.

Therefore, this newly added test is helpful to understand the effects of
the first Google Docs opening, by rebooting the DUT every time a single
iteration of DocsCUJ runs.

BUG=chromium:1328923
TEST=tast run device meta.ColdbootDocsCUJ*

Change-Id: I934cda4ce51f5c62913c129455ee761adf603834