Allow namespaced ARC tts system property
This introduces a temporary hack which allows two property names for the
same purpose. The hack will be in place until ARC repositories use the
ro.vendor.arc.* property name.

I tried to test this with
http://cs/rvc-arc/vendor/partner_gms/GmsCoreRelease/README.md but it
was especially involved I skipped the directions before the "Manual
tests" heading and got errors.

BUG=b:237255015
TEST=build

Change-Id: I3c07a05a79545cada710a577bbf11ffacbd4f06a