tast-tests: Add GetDeskCount function
Add a GetDeskCount function that calls into the
autotestPrivate.getDeskCount function and returns the number of open
desks in the DesksController. Also uses the new function to check that a
desk was actually deleted in wmp.OverviewCloseAll and
wmp.OverviewCombineDesks, and that a desk was not deleted in
wmp.OverviewCloseAllUndo.

BUG=chromium:1251558
TEST=tast -verbose run <IP> wmp.OverviewCloseAll, tast -verbose run <IP>
wmp.OverviewCloseAllUndo, tast -verbose run <IP>
wmp.OverviewCombineDesks

Change-Id: I7e9aa8ec5253725ca0b32b7438c952fdc3864ed9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3781076
Reviewed-by: Yanzhu Du <yzd@google.com>
Tested-by: Ben Becker <benbecker@chromium.org>
Commit-Queue: Ben Becker <benbecker@chromium.org>
Reviewed-by: Kyle Shimabukuro <kyleshima@chromium.org>