tast-test: Cleanup CWS install without using settings
Currently, we clean up the test by uninstalling an app using the
Settings app. When a test fails, the Settings app is opened and it could
affect following tests by giving preference to the Settings app instead
of the newest install.

Unlike ARC, the chrome web store does not have a workaround for
uninstalling apps.

For this reason, during cleanup we would try to open the launcher and
uninstall the app with the context menu.

BUG=chromium:1344547
TEST=tast run <DU>T launcher.RecentApps.cws*

Change-Id: Ib49914f5d4cf834ca1c5e979837c901388722b20