tast-test: Recreate the chrome session for RecentApps tests
One of the reasons why the RecentApps tests are failing, is because the
order of the recent apps is not in the expected order. This could be
because as we run other tests before these ones, they might have left
Recent apps in a state where a preferred app is shown first because it
was used in the previous test.

To account for these cases, it is best if the test logs in again to have
a clean state.

This CL also removes the previously defined fixture for Arc boot session
with Productivity launcher enabled, as the test will no longer use it.

BUG=chromium:1344547, b:232144821
TEST=tast run <DU>T launcher.RecentApps.*

Change-Id: Ib49914f5d4cf834ca1c5e979837c901388722b20