wifi: Implement a remote captive portal test with a fake HTTP server
This modification creates the HTTP server and DNS server on the
remote router. When the portal detection is enabled and connecting
to the remote router from the DUT, the DNS server will reroute HTTP
calls to the HTTP server, which will cause the service state to be
changed to redirect-found.

BUG=b:237285121
TEST=tast run wifi.CaptivePortalHTTP

Change-Id: Ia303c1bca4a00d2b4458a4eb65da977e9d2f4524