wifi: Implement a remote captive portal test with a fake HTTP server
This modification creates the HTTP server and DNS server on the
remote router. When the portal detection is enabled and connecting
to the remote router from the DUT, the DNS server will reroute HTTP
calls to the HTTP server, which will cause the service state to be
changed to redirect-found.

BUG=b:237285121
TEST=tast run wifi.CaptivePortalHTTP and run wifi_matfunc to check
that all tests pass

Change-Id: Ia303c1bca4a00d2b4458a4eb65da977e9d2f4524
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3782775
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Ting-Hao Lin <tinghaolin@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Ting-Hao Lin <tinghaolin@google.com>