tast-tests: syzkaller: fail test with unsupported syscalls
The tast test fails to catch a silent failure when an unsupported
syscall is enabled. As a result the fuzzer does not run,
but no errors are reported by the tast test, causing this failure to
go unnoticed.

Update the tast test to look for this error and fail the test if it
shows up.

BUG=b:239942103
TEST=./fast_build.sh -T # passes
TEST=tast -verbose run -var=syzkaller.IsLocal="true" localhost:2202 syzkaller.Wrapper.non_brya # ensure that test fails with an unsupported syscall.

Change-Id: I77ebff0d2c6027543092e454a9e109e6e6ab86c3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3782071
Tested-by: Zubin Mithra <zsm@chromium.org>
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Commit-Queue: Zubin Mithra <zsm@chromium.org>
Reviewed-by: Seewai Fu <seewaifu@google.com>