tast-tests: Add passthrough cmd decoder variants for a few tests
The passthrough command decoder is a Chrome GPU feature currently
being rolled out via Finch. Since this feature changes a lot, we'd
like to run a few simple end-to-end Tast tests as a sanity check
that changes to ANGLE don't cause crashes on CrOS.

Note that this is a temporary addition. Once we launch the feature
by default, we can remove these variants.

Bug: b:233238923
Change-Id: I23da16a06d45a9a606f9f50b2b3d0ef1ae05d303