[Nearby][WifiCreds] Use WaitForServiceProperties
In order to avoid not catching the Wi-Fi Network (i.e., service),
we switch FindMatchingService to WaitForServiceProperties in order
to avoid polling before the network is fully configured. Additionally,
this change will use some UI Automation to open the known Wi-Fi networks
page from the success screenshot, which would verify that the
notification shows up would provide more context in screenshots if the
tast test fails.

BUG=b:239434458
TEST=this

Change-Id: I7912126400d1a7944c8ab03daad32e3342b90ee4