arcappcompat: Update Minecraft education in tast
BUG=b:236431792
TEST=tast run <eve> arcappcompat.Minecraft.vm_tablet_mode
tast run <eve> arcappcompat.Minecraft.vm_clamshell_mode
tast run <kevin> arcappcompat.Minecraft.tablet_mode
tast run <kevin> arcappcompat.Minecraft.clamshell_mode

Minecraft is failing to launch on some of the Chromebooks
(kevin, bluebird, sarien, bobba360). It is observed that
UI flow is different on some devices. On some devices
“View EULA” button is clicked and it opens up a web page
in the chrome browser. Hence the “MinecraftEDU” android
app is in the background. Added a code to close a web
page in the chrome browser to make the “MinecraftEDU" android
app to be visible. The code helps to fix the “failed to
launch the app” error.

Change-Id: I243ea1b132e3e3d90debd6312e68ced54b0dc395