tast: Add test for text editing with VK
Add tast test to check that text editing with the VK works correctly in
Ash and Lacros. The test includes various typing actions after changing
focus between text fields and moving the cursor around.

Note that the test is currently failing, since typing in the middle of a
composition still has bugs on both Ash and Lacros.

BUG=b:237353274
TEST=tast run <DUT> inputs.VirtualKeyboardTextEditing*

Change-Id: I14ad5c607cc283c99c30d05c65fa3c78eeb1ec0e