crostini: rework on resize* tests
The current resize* tests do not do what is expected for resizing,
as the helper functions for resizing are problematic: the calculation
of max size and min size is unstable and the tests may use the same
value as both max and min. Also the function to change value on slider
is unstable as it stops if two consecutive disk sizes checks returns
the same value, which could happen unexpectedly due to latency issues.

This CL makes the following changes:
1, Remove the direction input param, which is trivial.
2, Read the max size and min size directly from the node info.
3, Rewrite the changingSize function to solely rely on the target size.
4, Revise the affected tests accordingly.

BUG=b:239368868
TEST=tast run <DUT> crostini.Resize*

Change-Id: I0ba45a9d469930a4f914fde82e7ee9dc8ea09926