tast-tests: Add a betty variant for enterprise.ARCProvisioning
Currently presubmit tests for vendor/partner_gms are running unstable
version of the test without retries.

This causes the test to fail for unrelated reasons on betty. By adding a
new betty variant we can update presubmit config to run these tests that
are doing retries but are still excluded from critical category.

BUG=b:186274474
TEST=CQ

Change-Id: I49482426cc96c3ecf36806bd56121962a5ec33fd