ui: Fix locating Meet window
Reading the window title is racy depending on if the tab has fully
loaded before we try and find it. If the tab hasn't been fully loaded
the window title will be meet.google.com. This change makes it so we
correctly locate the window in either case.

BUG=b:236001108
TEST=tast run 192.168.0.15 ui.MeetCUJ.2p_enterprise

Change-Id: Ieac5db9857b4f4fec62c113407e0e9f0f1e5abf9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3787193
Auto-Submit: Raul Rangel <rrangel@chromium.org>
Tested-by: Raul Rangel <rrangel@chromium.org>
Commit-Queue: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Avery Musbach <amusbach@chromium.org>