fusebox: make tast tests call terminateDaemonIfNeeded
The fusebox tast tests run fine on the CQ bots, but on VM's the tests can fail with a fusebox daemon D-BUS ownership error [1].

Suspect that a prior test, perhaps a Chrome test, has failed to clean
up and left an instance of the /usr/bin/fusebox daemon running.

Add terminateDaemonIfNeeded helper to mitigate: call it test setup to
ensure no pre-existing daemon instance is running.

[1] See https://crbug.com/issue=1329684#c28

BUG=chromium:1329684
TEST=tast run root@DUT platform.Fusebox*

Change-Id: I7d81c86798714fe7c4dee895805656dcb0279f62