rpcdut: Check if DUT is not going to reboot after update
After FPMCU firmware update during boot, reboot is requested to
initialize FPMCU properly. On some DUT (e.g. hatch) the SSH connection
can be established between firmware update and second reboot (it
depends on many factors like DHCP server), so as a result connection to
DUT will be broken eventually.

To make sure that ChromeOS is not going to reboot we can check the state
of the 'boot-firmware-update' and 'reboot' services. The former won't be
marked as started until update finished, the latter won't be stopped if
DUT is going to reboot (runlevel changed to 6)

BUG=b:239425260
TEST=Run 'tast run <IP> firmware.FpUpdater' on zork, hatch, nami
     nocturne, strongbad
TEST=TAST unit tests passed.

Change-Id: I6582ddf7971c08c5040b0ef57608b1f1f50d3dfc