crostini: increase timeout for PostTest
The fixture restarts Crostini in PostTest  to clean up if a test fails.

The original timeout (30s) is not sufficient on some devices if
restartCrostini is needed .

i.e. the fixture fails to restart Crostini after
crostini.AppGeditFilesharing.clamshell_stable fails and it leads to the
failure of AppGeditUnshareFolder in https://ci.chromium.org/ui/p/chrome/builders/ci/chromeos-eve-chrome/23673/overview

BUG=b:231258823
TEST=tast run <DUT> crostini.AppsGedit*

Change-Id: I92678bbd702855e75c7b74ab2a2ab16a4775fb9f