usbprintertests: Refactor RunPrintTest
RunPrintTest sets up a printer, runs a print job, and verifies the
result.  Split out the print job portion so that tests that need a
custom printer setup don't have to copy and paste the rest of the
behavior.  In the process, refactor the series of string parameters into
a struct to make it easier to see what's being passed in.

BUG=b:238353195
TEST=tast run printer.Print*

Change-Id: I4d20381a56b4896e18e578ac4c83ae6f2c8fdd6b