tape: Add a sample test to confirm TAPE functionality
Add a test to confirm TAPE is able to lease generic accounts when run
locally, and in the lab. This only needs to run on a single board since
it serves as a simple check to confirm DUTs can reach the TAPE service.

BUG=chromium:1339418
TEST=tast run --var tape.local_refresh_token=<VALUE>
tape.GenericAccount*

Change-Id: I5f36b454406f31562ea0bda0b24ffbb9a1dbe541
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3788525
Commit-Queue: David Welling <davidwelling@google.com>
Tested-by: David Welling <davidwelling@google.com>
Reviewed-by: Alexander Hartl <alexanderhartl@google.com>