tast-tests: refactor for lacros extensions
Currently ash and lacros-chrome are loading custom extensions in Tast
in a different way that might be confusing to test writers.

This change provides a more consistent way of loading lacros extensions
similarly to ash extensions.

The steps to enable lacros extensions are as follows:
1. Create temp ext dirs with test content in tests or fixtures
2. Get a chrome option from chrome.LacrosUnpackedExtension(extDir)
3. Pass it to chrome.New or browserfixt.SetUpWithNewChrome
4. Then let chrome take care of the rest of the work

Also it adds new lacros fixtures that load fake apps from lacros
extensions. The existing fixtures that use the old extension fixtures
will be updated in a follow up CL.

See go/lacros-tast-extensions for details.

BUG=chromium:1343713, chromium:1309565
TEST=tast run '("group:mainline" && "dep:lacros" && !informational)'  // all critical
TEST=tast run shelf.AppRearrangement.rearrange_chrome_apps  // tests using ash extensions

Change-Id: I1416cf074051f3fa2b737318e4e9cffdc98863fe