lacros.Migrate: Fix history detection
This is another attempt to fix the frequent test failure where
a browsing history entry is not found (during the profile preparation
phase).

The previous attempt unfortunately made the problem worse because the
history page's documentstate becomes 'ready' even when page hasn't fully
loaded yet.

The CL tries to do the simplest thing possible: wait for the item to
show up, and if it doesn't then reload once and wait again (this time
longer). We know for a fact that sometimes a reload is necessary but we
should never need more than one. We also know for a fact that the
standard timeout of 15s is not always enough, hence the second wait uses
twice that.

BUG=chromium:1330419,chromium:1340820
TEST=tast run $DUT lacros.Migrate.*

Change-Id: Ia927a159e39bbe75d8f15df45e6922a279d38b21