tast-tests: remove cryptohome.UserSecretStashAddPin test
These PIN tests are already covered (more completely) by the tests in
hwsec.PinWeaverWithAuthAPI, so this is just redundant.

I think you could make a good argument that actually the
PinWeaverWithAuthAPI tests are really cryptohome tests and not hwsec
tests, since they're exercising the APIs and the cryptohome backends,
but I think that's a separate argument. Regardless of where those tests
belong I don't think it makes sense to have a separate pile that covers
the same ground.

BUG=b:240316365
TEST=tast run <dut> cryptohome.*

Change-Id: I6ee89ee69a41254d5a288b05a00b39c3e9bdfc01