printpreview: Account for tablet-mode keyboard
The default control in the destination dialog is an edit box.  In tablet
mode, this means that the onscreen keyboard will pop up immediately when
the destination dialog is opened.  This covers the particular printer
entry that the test needs to select, so we need to close the keyboard
before trying to click the printer.  We do this by clicking the dialog
title instead of calling vkb.HideVirtualKeyboard because the
documentation indicates that HideVirtualKeyboard is only intended to be
used during cleanup.  Clicking on the dialog title is harmless when not
in tablet mode, so we can do it unconditionally.

This also fixes a bug where the destination dropdown is not found if an
Android activity restart left a printer selected.

BUG=b:238736276
TEST=tast run arc.Print with and without tablet mode

Change-Id: Ic7652b84d28e46174d93de5a8ed7159fb05ba1fb