health: Move usb utils to shared package
We want to add a new non-healthd test that needs to do the same USB
device parsing as health.usb.ExpectedDevices.  Move this package to be
shared so that the new test can access it.  The shared package is called
usbutil instead of usb to match the naming of similar existing packages.

BUG=b:238353195
TEST=tast run health.*

Disallow-Recycled-Builds: test-failures
Change-Id: Id24207de0241a8fd071d094f649fa083aa17d90c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3803505
Reviewed-by: Wei-Luan Wang <weiluanwang@google.com>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Chung-Sheng Wu <chungsheng@google.com>
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>