tast-tests: Update for handling WV14 test variants
This adds support for running with either WV14 or WV16 compatible
firmware on Intel devices. WV16 is configured as the default, and if we
detect a version incompatibility then we switch to the WV14 variant.
This will be removed once all volteer/brya devices have WV16 FW pinned.

BUG=b:240453750
TEST=tast run DUT video.CDMOEMCrypto.* with WV14 and WV16 firmware

Cq-Depend: chrome-internal:4900345
Change-Id: I408828516f5426b7273d39070489863b8235e63a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3789935
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Avery Musbach <amusbach@chromium.org>
Commit-Queue: Jeffrey Kardatzke <jkardatzke@google.com>