Including Basic Goole Meet CUJs in crosbolt nightly tests.
Adding the Basic Meet CUJs to the list of nightly tests. At this stage,
only Baisc tests are added to check the stability of framework. If it
ran smoothly Plus and Premium tests will be added to. Becuase the
device list for the tests are shared between local and remote tests, it
moved to the common directory and importt changed accordingly.

BUG=None
TEST=tast run -build=True <DUT_IP>
ui.GoogleMeetCUJ.basic_two_crosbolt

Change-Id: Ifce55e814897006adbfa3e252ffff8d063370ede