tast-tests: Add more complete Chameleond XMLRPC API
This change adds a series of interfaces that reflect Chameleond
XMLRPC methods as well as base implementations of each interface.

All known RPC methods of the base ChameleondInterface and bluetooth
devices have been added. Some more complex arguments and return values
are managed through structs for ease of use. Key Chameleond constants
that are used by these methods, such as PortID have been included as
well.

All the Chameleond XMLRPC calls are made using a new CallBuilder
utility. This significantly reduces the complexity of each call, with
most implemented with a single line of code.

The method docs for the XMLRPC calls have been copied from the
corresponding methods in the Chameleond codebase with only minor
wording changes. As this is only a client API, it is not expected to
include every detail regarding the methods' functionality. For complete
and current documentation for each method, users are directed to the
Chameleond codebase.

All Chameleond XMLRPC interfaces expose the base RPC method to the
user so that methods not yet included in this client API may still be
used with minimal effort.

BUG=b:217633791
TEST=tast unit tests passed, including the new ones that were added.
Also passed assistant.OpenSettingsAudioChameleon using a Chameleon
with an audio board.

Change-Id: I06738474ab7bc315e5e34064ba93194c5068d123
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3790144
Reviewed-by: Alexandru Stan <amstan@chromium.org>
Tested-by: Jared Bennett <jaredbennett@google.com>
Commit-Queue: Jared Bennett <jaredbennett@google.com>
Reviewed-by: Jonathan Fan <jonfan@chromium.org>
Reviewed-by: Patrick Meiring <meiring@google.com>