lacrosinfo: Replace Running with State
Now that autotest_private's LacrosInfo structure contains the Lacros
state, we can use that instead of the boolean Running.

In fact, we can improve Lacros's Close() by waiting for the state
to become "Stopped" instead of merely becoming something other than
"Running". This might reduce the number of flakes due to
crbug.com/1316237.

Finally, remove a spurious use of PollBreak.

BUG=None
TEST=fast_build.sh -T
TEST=tast run $DUT lacros.Basic

Change-Id: I3cb6d25015ef8cf3280673ce983ea3d39a0dbcb9