lacrosproc.Root: Eliminate LacrosLocation parameter
The parameter made the function inconvenient to use. Remove
it in favor of fetching the path of Lacros via lacrosinfo.

This also fixes the function for the case of deployed Lacros,
where an incorrect path was used.

(The only test currently using the function is unfortunately broken.
I've filed an issue here: crbug.com/1348073)

BUG=None
TEST=fast_build.sh -T

Change-Id: Ie558219b400aa3527765d64302ba5a6e7628d90d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3788958
Commit-Queue: Georg Neis <neis@chromium.org>
Reviewed-by: Eliot Courtney <edcourtney@chromium.org>
Tested-by: Georg Neis <neis@chromium.org>