tast-tests: Add a trailing slash to ui.PerftestURL
There has been an issue with browserfixt not opening some URLs for
lacros without a trailing slash added at the end in Tast. It is because
the URL should match exactly what it gets loaded in the tab.
This changes ui.PerftestURL to "chrome://version/" to make it work
for both ash and lacros-chrome.

We can do the same for some other URLs in a follow up.

BUG=b:225229931
TEST=Run some tests that open ui.PerftestURL at least for each package
TEST=tast run ui.WindowCyclePerf* ui.VideoCUJ* ui.HotseatAnimation* ui.WindowResizePerf*
TEST=tast run shelf.AutoHideSmoke*
TEST=tast run assistant.EmbeddedUIBubbleLauncherOpenAndCloseAnimationPerf*

Change-Id: Ic508ebabccbcf30bb16f1c7dc256dad64df6443c